Fixes #44 Updated the prevalidate request to remove the dimension intervals from the request.  Moved tx_utils to the root of the project as a general utils module.
